pyproject.toml
[project]
name = "pubmed-scraper"
version = "0.1.0"
description = "A Python tool for scraping article metadata from PubMed using the NCBI E-utilities API"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "anthropic>=0.18.0",
]

[project.scripts]
pubmed-scraper = "pubmed_scraper:main"
